llama-index/requirements.txt
llama-index-core>=0.11.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-readers-file>=0.3.0
qdrant-client>=1.9.0
uagents>=0.23.0
uagents-core>=0.4.0
python-dotenv>=1.0.0
cosmpy
requests>=2.31.0
openai>=1.0.0
